File reorganisation
In preparation for turning the visualisation components in `./src/lib`
into a npm package, while the other files in `./src` are for testing /
developing the published components. For this reason all exported
components / functions etc are exposed in `./lib/index.js`.
