+++ src/App.js
@@ -1,9 +1,7 @@
 import React from 'react';
 import { Container } from './Container.js';
-import { useDataFetch } from './dataFetch.js';
-import { Panels } from "./Panels.js"
-import {Status} from "./Status";
-import { ErrorBoundary } from './ErrorBoundary.js';
+import { Panels } from './Panels.js';
+import { useDataFetch, Status, ErrorBoundary } from './lib/index.js';
 import './fonts.css';
 
 function App() {
