+++ src/lib/components/SmallMultiple.js
