+++ src/lib/index.js
@@ -1,0 +1,7 @@
+import { useDataFetch } from './utils/dataFetch.js';
+import { SmallMultiple } from './components/SmallMultiple.js';
+import { Status } from './components/Status.js';
+import { ErrorBoundary } from './components/ErrorBoundary.js';
+import { useLegend } from './utils/Legend.js';
+
+export {useDataFetch, useLegend, SmallMultiple, Status, ErrorBoundary};
