+++ src/lib/utils/Legend.js
